requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
sqlalchemy
mysql-connector-python
python-dotenv
yfinance
